Confidence sets for ranks

Given estimates and their covariance matrix of a certain feature for a set of populations, calculate confidence sets for the ranks of populations, where populations are ranked by the feature values.

csranks(
  x,
  Sigma,
  coverage = 0.95,
  cstype = "two-sided",
  stepdown = TRUE,
  R = 1000,
  simul = TRUE,
  indices = NA,
  na.rm = FALSE,
  seed = NA
)

Arguments

x

vector of estimates.

Sigma

covariance matrix of x. Note, that it must be covariance matrix of feature means, not features themselves.

coverage

nominal coverage of the confidence set. Default is 0.95.

cstype

type of confidence set (two-sided, upper, lower). Default is two-sided.

stepdown

logical; if TRUE (default), stepwise procedure is used, otherwise single step procedure is used. See Details section for more.

R

number of bootstrap replications. Default is 1000.

simul

logical; if TRUE (default), then simultaneous confidence sets are computed, which jointly cover all populations indicated by indices. Otherwise, for each population indicated in indices a marginal confidence set is computed.

indices

vector of indices of x for whose ranks the confidence sets are computed. indices=NA (default) means computation for all ranks.

na.rm

logical; if TRUE, then NA's are removed from x and Sigma (if any).

seed

seed for bootstrap random variable draws. If set to NA (default), then seed is not set.

Value

A csranks object, which is a list with three items:

L

Lower bounds of the confidence sets for ranks indicated in indices

rank

Raw rank estimates using irank with default parameters

U

Upper bounds of the confidence sets.

Details

IMPORTANT: make sure, that the Sigma is a (perhaps estimated) covariance matrix of estimates of feature means across populations, not of features themselves. For example, sample of size \(n\) of a feature following a standard normal distribution has variance \(\sigma^2=1\), but mean from such sample has variance \(1/n\). We refer to the latter.

The command implements the procedure for construction of confidence sets for ranks described in the referenced paper below. Generally, it consists of verification of a large set of hypotheses. After rejection of certain set of hypotheses, one can terminate the procedure or keep verifying a smaller set of hypotheses that were not rejected so far. The former corresponds to stepdown=FALSE; the latter to stepdown=TRUE.

From a practical point of view, stepdown=TRUE takes more time, but usually results in tighter (better) confidence sets.

Parametric bootstrap used to calculate distribution for confidence sets based on the multivariate normal distribution.

References

Mogstad, Romano, Shaikh, and Wilhelm (2023), "Inference for Ranks with Applications to Mobility across Neighborhoods and Academic Achievements across Countries", forthcoming at Review of Economic Studies

pdf link doi:10.1093/restud/rdad006

Examples

# Setup example data
n <- 10
x <- seq(1, 3, length = n)
Sigma <- matrix(0.001, nrow = n, ncol = n)
diag(Sigma) <- 0.04

# Run csranks to get confidence sets for ranks of features
csranks(x, Sigma)
#> $L
#>  [1] 7 6 5 4 3 2 1 1 1 1
#> 
#> $rank
#>  [1] 10  9  8  7  6  5  4  3  2  1
#> 
#> $U
#>  [1] 10 10 10 10  9  8  7  6  5  4
#> 
#> attr(,"class")
#> [1] "csranks"

# If you assume that the feature measurements are independent 
# (or have access only to variances / standard errors estimates),
# then pass a diagonal covariance matrix.
Sigma <- diag(rep(0.04, 10))
csranks(x, Sigma)
#> $L
#>  [1] 6 5 4 3 2 1 1 1 1 1
#> 
#> $rank
#>  [1] 10  9  8  7  6  5  4  3  2  1
#> 
#> $U
#>  [1] 10 10 10 10 10  9  8  7  6  5
#> 
#> attr(,"class")
#> [1] "csranks"