Simultaneous confidence sets for ranks using the bootstrap

This function is called by csranks_simul and by csranks_multinom when simul=TRUE and bootstrap=TRUE.

csranks_simul_boot(x, param, coverage = 0.95, cstype = "two-sided",
  stepdown = TRUE, bootdist = "normal", R = 1000, indices = NA,
  seed = NA)

Arguments

x
vector of estimates

param
parameters required for bootstrap: if bootdist="normal", then param is a vector of standard errors of x; if bootdist="multinom", then param is number of trials.

coverage
nominal coverage of the confidence set. Default is 0.95.

cstype
type of confidence set (two-sided, upper, lower). Default is two-sided.

stepdown
logical; if TRUE (default), stepwise procedure is used, otherwise single step procedure is used.

bootdist
distribution to be used for parametric bootstrap; either normal (default) or multinomial.

R
number of bootstrap replications. Default is 1000.

indices
vector of indices of x for whose ranks the confidence sets are computed. indices=NA (default) means computation for all ranks.

seed
seed for bootstrap random variable draws. If set to NA (default), then seed is not set.

Value

L, U lower and upper bounds of the confidence set for ranks indicated in indices.