Linear model for ranks

Fit a linear model with a single rank response and a single rank regressor (and possibly other usual regressors).

lmranks(
  formula,
  data,
  subset,
  weights,
  na.action = stats::na.fail,
  method = "qr",
  model = TRUE,
  x = FALSE,
  qr = TRUE,
  y = FALSE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset = offset,
  omega = 1,
  na.rm = FALSE,
  ...
)

# S3 method for lmranks
plot(x, which = 1, ...)

# S3 method for lmranks
predict(object, newdata, ...)

# S3 method for lmranks
vcov(object, complete = TRUE, ...)

Arguments

formula

An object of class "formula": a symbolic description of the model to be fitted. Exactly like the formula for linear model except that rank terms, r(), can be added to specify that the linear regressor depends on ranks of regressors and to specify a rank response. See Details.

data

an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.

subset

currently not supported.

weights

currently not supported.

na.action

currently not supported. User is expected to handle NA values on his own.

method

the method to be used; for fitting, currently only method = "qr" is supported; method = "model.frame" returns the model frame (the same as with model = TRUE, see below).

model, y, qr

logicals. If TRUE the corresponding components of the fit (the model frame, the response, the QR decomposition) are returned.

x

For lmranks: Logical. Should model matrix be returned?

For plot method: An lmranks object.

singular.ok

logical. If FALSE (the default in S but not in R) a singular fit is an error.

contrasts

an optional list. See the contrasts.arg of model.matrix.default.

offset

this can be used to specify an a priori known component to be included in the linear predictor during fitting. This should be NULL or a numeric vector or matrix of extents matching those of the response. One or more offset terms can be included in the formula instead or as well, and if more than one are specified their sum is used. See model.offset.

omega

as in frank.

na.rm

If FALSE, raises an error is any NA values in ranked regressors or response are encountered. If TRUE, ranks for non-NA entries are calculated by ignoring NA values. For NA values, NA ranks are returned and handled later by na.action.

...

For lm(): additional arguments to be passed to the low level regression fitting functions (see below).

which

As in plot.lm. Currently only no.1 is available.

object

lmranks object.

newdata

An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

complete

logical indicating if the full variance-covariance matrix should be returned also in case of an over-determined system where some coefficients are undefined and coef(.) contains NAs correspondingly. When complete = TRUE, vcov() is compatible with coef() also in this singular case.

Value

An object of class lmranks, inheriting (as well as possible) from class lm. See the lm documentation for more.

Additionally, it has an omega entry, corresponding to omega argument, and a rank_terms_indices - an integer vector with indices of entries of terms.labels attribute of terms(formula), which correspond to ranked regressors.

A number of methods defined for lm does not yield theoretically correct results when applied to lmranks objects; errors or warnings are raised consciously. Also, the df.residual component is set to NA, since the notion of effects of freedom for the rank models is not theoretically established.

Details

This function is useful in case when relationship not between variables themselves, but their rank (or, put differently, their ECDF value) is of interest. The variables to be ranked (both dependent and independent) can be marked with r().

The r() is a private alias for frank with fixed increasing argument. omega argument may be specified globally (as a specification of rank definition) in the call to lmranks.

As a consequence of the order, in which model.frame applies operations, subset and na.action would be applied after evaluation of r(). In such a case, returned coefficients and standard errors might no longer be correct. The user must handle the NA values and may filter the data on his own.

Currently, only models at most one rank regressor are available. The single response might be either ranked or continuous.

Many functions defined for lm also work correctly with lmranks. This includes coef, model.frame, model.matrix, resid and update. On the other hand, some would not return correct results.

Methods (by generic)

plot(lmranks): Plot diagnostics for an lmranks object

Displays plots useful for assessing quality of model fit. Currently, only one plot is available, which plots fitted values against residuals (for homoscedacity check).

predict(lmranks): Predict method for Linear Model for Ranks Fits

vcov(lmranks): Calculate Variance-Covariance Matrix for a Fitted lmranks object

Returns the variance-covariance matrix of the regression coefficients (main parameters) of a fitted lmranks object.

Warning

Wrapping r() with other functions (like log(r(x))) will not recognize correctly the mark (because it will not be caught in terms(formula, specials = "r")). The ranks will be calculated correctly, but their transformation will be treated later in lm as a regular regressor. This means, that the corresponding regression coefficient will be calculated correctly, but the standard errors, statistics etc. will not.

r, .r_predict and .r_cache are special expressions, used internally to interpret r mark correctly. Do not use them in formula.

See also

Examples

Y <- c(3,1,2,4,5)
y_frank <- c(0.6, 1.0, 0.8, 0.4, 0.2)
X <- 1:5
omega <- 0.5
x_frank <- c(1.0, 0.8, 0.6, 0.4, 0.2)
W <- matrix(y_frank * 0.1 + 5 + rnorm(5, sd = 0.1), ncol = 1)

lmranks(r(Y) ~ r(X) + W)
#> 
#> Call:
#> lmranks(formula = r(Y) ~ r(X) + W)
#> 
#> Coefficients:
#> (Intercept)         r(X)            W  
#>     12.9730       0.7977      -2.5422  
#> 
# naive version with same regression coefficients, but incorrect 
# standard errors and statistics:
lm(y_frank ~ x_frank + W)
#> 
#> Call:
#> lm(formula = y_frank ~ x_frank + W)
#> 
#> Coefficients:
#> (Intercept)      x_frank            W  
#>    -12.7303       0.7977       2.5422  
#> 

data(mtcars)
lmranks(r(mpg) ~ r(hp) + ., data = mtcars)
#> 
#> Call:
#> lmranks(formula = r(mpg) ~ r(hp) + ., data = mtcars)
#> 
#> Coefficients:
#> (Intercept)        r(hp)          cyl         disp           hp         drat  
#>    0.489054     0.094401    -0.047335     0.000655    -0.001717    -0.012264  
#>          wt         qsec           vs           am         gear         carb  
#>   -0.152140     0.036960    -0.038424     0.057301     0.072761    -0.008447  
#>