Confidence sets for ranks

Confidence sets for ranks

csranks(x, sd, coverage = 0.95, cstype = "two-sided",
  stepdown = TRUE, R = 1000, simul = TRUE, indices = NA,
  na.rm = FALSE, seed = NA)

Arguments

x
vector of estimates

sd
vector of standard errors of x

coverage
nominal coverage of the confidence set. Default is 0.95.

cstype
type of confidence set (two-sided, upper, lower). Default is two-sided.

stepdown
logical; if TRUE (default), stepwise procedure is used, otherwise single step procedure is used.

R
number of bootstrap replications. Default is 1000.

simul
logical; if TRUE (default), then simultaneous confidence sets are computed, which jointly cover all populations indicated by indices. Otherwise, for each population indicated in indices a marginal confidence set is computed.

indices
vector of indices of x for whose ranks the confidence sets are computed. indices=NA (default) means computation for all ranks.

na.rm
logical; if TRUE, then NA's are removed from x and sd (if any).

seed
seed for bootstrap random variable draws. If set to NA (default), then seed is not set.

Value

L, U lower and upper bounds of the confidence set for ranks indicated in indices.

Details

Implentation of the confidence sets proposed in Mogstad, Romano, Shaikh, and Wilhelm ("Inference for Ranks with Applications to Mobility across Neighborhoods and Academic Achievements across Countries", CeMMAP Working Paper CWP10/20).

Examples

x <- seq(1,3,length=10)
sd <- rep(0.2,10)
csranks(x, sd)

#> $L
#>  [1] 7 6 5 4 3 2 1 1 1 1
#> 
#> $U
#>  [1] 10 10 10 10  9  8  7  6  5  4
#>