Compute ranks from feature values

Given estimates of a certain feature for a set of populations, calculate the integer ranks of populations, i.e. places in ranking done by feature values. The larger (or smaller) feature value, the higher the place and the lower the integer rank (lowest, 1, is the best place).

irank(x, omega = 0, increasing = FALSE, na.rm = FALSE)

frank(x, omega = 0, increasing = FALSE, na.rm = FALSE)

Arguments

x

vector of values to be ranked

omega

numeric; numeric value in [0,1], each corresponding to a different definition of the rank; default is 0. See Details.

increasing

logical; if TRUE, then large elements in x receive a large rank. In other words, larger values in x are lower in ranking. Otherwise, large elements receive small ranks.

na.rm

logical; if TRUE, then NA's are removed from x. In other case the output for NAs is NA and for other values it's for extreme possibilities that NA values are actually in first or last positions of ranking.

Value

Integer vector of the same length as x containing the ranks.

Details

omega (\(\omega\)) value determines, how equal entries in x should be ranked; in other words how to handle ex aequo cases. If there are none, then the parameter does not affect the output of this function. For example, let's say, that \(n\) largest entries in x are equal. Those entries could receive (minimum) rank 1 or (maximum) rank \(n\) or some value in between.

Suppose, that we want to assign rank to \(n\) equal values in an array. Denote their minimum rank as \(r\) and maximum as \(R = r + n - 1\). Then the assigned rank is an average of minimum and maximum rank, weighted by \(\omega\): $$r(1-\omega) + R\omega$$

Functions

frank(): Compute fractional ranks

This method returns ranks in form of fractions from [0-1] interval. Smaller values (closer to 0) indicate higher rank.

Examples

irank(c(4,3,1,10,7))
#> [1] 3 4 5 1 2
irank(c(4,3,1,10,7), omega=1) # equal to previous ranks because there are no ties
#> [1] 3 4 5 1 2
irank(c(4,3,1,10,7), omega=0.5) # equal to previous ranks because there are no ties
#> [1] 3 4 5 1 2
irank(c(4,4,4,3,1,10,7,7))
#> [1] 4 4 4 7 8 1 2 2
irank(c(4,4,4,3,1,10,7,7), omega=1)
#> [1] 6 6 6 7 8 1 3 3
irank(c(4,4,4,3,1,10,7,7), omega=0.5) 
#> [1] 5.0 5.0 5.0 7.0 8.0 1.0 2.5 2.5
frank(c(4,3,1,10,7))
#> [1] 0.6 0.8 1.0 0.2 0.4
frank(c(4,3,1,10,7), omega=1) # equal to previous ranks because there are no ties
#> [1] 0.6 0.8 1.0 0.2 0.4
frank(c(4,3,1,10,7), omega=0.5) # mid-ranks, equal to previous ranks because there are no ties
#> [1] 0.6 0.8 1.0 0.2 0.4
frank(c(4,4,4,3,1,10,7,7))
#> [1] 0.500 0.500 0.500 0.875 1.000 0.125 0.250 0.250
frank(c(4,4,4,3,1,10,7,7), omega=1)
#> [1] 0.750 0.750 0.750 0.875 1.000 0.125 0.375 0.375
frank(c(4,4,4,3,1,10,7,7), omega=0.5) # mid-ranks
#> [1] 0.6250 0.6250 0.6250 0.8750 1.0000 0.1250 0.3125 0.3125